+++ src/runtime/components/NuxtTime.vue
@@ -1,5 +1,6 @@
 <script setup lang='ts'>
 import { computed, getCurrentInstance, useNuxtApp, useHead } from '#imports'
+import script from '../../script.min.js';
 
 const props = withDefaults(defineProps<{
   locale?: string
@@ -61,20 +62,7 @@ if (process.server) {
     script: [{
       tagPosition: 'bodyClose',
       key: 'nuxt-time',
-      innerHTML: `
-        document.querySelectorAll('[data-n-time]').forEach(el => {
-          const date = new Date(el.getAttribute('datetime'));
-          const options = {};
-          for (const name of el.getAttributeNames()) {
-            if (name.startsWith('data-')) {
-              options[name.slice(5)] = el.getAttribute(name);
-            }
-          }
-
-          const formatter = new Intl.DateTimeFormat(options.locale, options);
-          el.textContent = formatter.format(date)
-        })
-      `.replace(/\s+/g, ' ')
+      innerHTML: script
     }]
   })
 }
